backend/requirements.txt
# AudioLetra Backend Dependencies
# Core Framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# LLM Integration
openai>=1.0.0
requests>=2.31.0

# Document Generation
python-docx>=0.8.11
reportlab>=4.0.0

# Template Engine (already included with Flask)
Jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-flask>=1.2.0

# Development Tools
python-dotenv>=1.0.0

# Utilities
uuid
datetime
typing
dataclasses
